api_v1/src/routers/patterns.py
# File: src/routers/patterns.py

from fastapi import APIRouter, HTTPException
from typing import Dict, Any

router = APIRouter()

@router.get("/")
async def get_learning_patterns():
    """
    Retrieve general learning pattern information
    """
    return {
        "message": "Learning Patterns Endpoint",
        "status": "Active"
    }

@router.post("/analyze")
async def analyze_learning_pattern(pattern_data: Dict[str, Any]):
    """
    Endpoint to analyze a specific learning pattern
    """
    try:
        # Placeholder for pattern analysis logic
        return {
            "pattern_input": pattern_data,
            "analysis_result": "Pattern analysis to be implemented"
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))