api_v1/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==43.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.25.1
langchain-core==0.3.31
langgraph==0.2.66
langgraph-api==0.0.16
langgraph-checkpoint==2.0.10
langgraph-cli==0.1.68
langgraph-sdk==0.1.51
langsmith==0.2.11
msgpack==1.1.0
orjson==3.10.15
packaging==24.2
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.45.2
structlog==24.4.0
tenacity==8.5.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
